+++ arrayjit/lib/cuda_backend_impl.mli
@@ -1,1 +1,3 @@
-module Fresh : functor () -> Ir.Backend_impl.Lowered_backend
+module Fresh (_ : sig
+  val config : Ir.Backend_intf.config
+end) : Ir.Backend_impl.Lowered_backend