+++ arrayjit/lib/metal_backend_impl.missing.ml
@@ -1,3 +1,8 @@
-module Fresh () = struct
+module Fresh (Config : sig
+  val config : Ir.Backend_intf.config
+end) =
+struct
+  let _ = ignore Config.config
+
   include Lowered_backend_missing
 end